requirements.txt
appdirs==1.4.3
asgiref==3.3.4
beautifulsoup4==4.9.3
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.4
django-channels==0.7.0
django-compressor==2.4.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-devserver==0.8.0
django-environ==0.4.5
django-heroku==0.3.1
django-htmlmin==0.11.0
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==2.2.0
django-rest-framework==0.1.0
django-shortcuts==1.6
django-social-auth==0.7.28
django-storages==1.11.1
djangorestframework==3.12.4
entrypoints==0.3
environ==1.0
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
httplib2==0.14.0
importlib-metadata==1.5.0
Markdown==3.3.4
markdown2==2.4.0
numpy==1.20.2
oauth2==1.9.0.post1
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.2.4
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.8.6
pycairo==1.16.2
pyDatalog==0.17.1
Pygments==2.2.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.2.0
pyRFC3339==1.1
pytz==2019.1
pyxdg==0.26
rcssmin==1.0.6
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
systemd-python==234
transliterate==1.10.2
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
whitenoise==5.2.0
wheel==0.36.2
